refactor(app): expose app.Run entry; cmd/agentdock/app.go is thin
App lifecycle logic now lives in app/app.go (both redis mode and
inmem mode). cmd/agentdock/app.go is a ~30-line cobra wrapper that
delegates to app.Run and propagates Version/Commit/Date link-time
vars into the app package so startup logs still report the build.

app/go.mod temporarily requires+replaces the worker module so app
can still assemble the inmem LocalAdapter. Phase 4 Task 34 moves
that assembly to the cmd layer and removes the app → worker edge.

cmd/agentdock/adapters.go (slackPosterAdapter) moves to app/
since it is only used from within Run.
