+++ app/adapters.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"log/slog"
